+++ Sources/FOSTestingVapor/LocalizableTestCase.swift
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#if canImport(Vapor)
+#if DEBUG && canImport(Vapor)
 import FOSFoundation
 import FOSMVVM
 import FOSTesting